feat: Adding README for Quicker React project
This commit introduces README.md for the Quicker React project. The file includes sections for Description, Installation, Usage, Creating a new component, Example, Contributing, Contact details, and Licence. The purpose is to provide concise project details including how to install and use the tool, contribute to the project and to stay in contact with the developer thereby improving the overall accessibility of the project.